fix warning on file.managed without changes
[WARNING ] State for file: /etc/apt/preferences - Neither 'source' nor
'contents' nor 'contents_pillar' nor 'contents_grains' was defined, yet
'replace' was set to 'True'. As there is no source to replace the file
with, 'replace' has been set to 'False' to avoid reading the file
unnecessarily.

Now, if file.managed, but when content is not changed (no source,
contents, ...), avoid the warning.