fix(apt_conf): pass context variables to the template as JSON object
Previously when passing the apt_conf context variable to the template
the variables would not necessarily retain their correct type. An
example of this would be when using a pillar value of `null` it would be
passed to the template as the string `"None"` instead of type `None`.

If we pass the apt_conf context variables to the template using a json
object the variables retain their correct types.